% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_report.R
\name{paged_report}
\alias{paged_report}
\title{Function for CCCM paged template}
\usage{
paged_report(other_css = NULL, number_sections = FALSE, ...)
}
\arguments{
\item{other_css}{Add extra css}

\item{number_sections}{Number section headings}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown report
}
\description{
Function for CCCM paged template
}
